% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.text}
\alias{radiant.text}
\alias{radiant.text-package}
\title{radiant.text}
\usage{
radiant.text()
}
\description{
radiant.text

Launch radiant.text in the default browser
}
\details{
See \url{http://radiant-rstats.github.io/docs} for documentation and tutorials
}

